pyproject.toml
[project]
name = "llm-ops-sandbox"
version = "0.1.0"
description = "Minimal observable LLM serving sandbox with FastAPI, Prometheus and Docker Compose."
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "prometheus-client>=0.20.0",
  "pydantic-settings>=2.4.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.setuptools.packages.find]
include = ["app*"]
